MULTIARCH-4545: Move LogVerbosity and SingleResourceObjectName
LogVerbosity can be considered a stable type that we won't change between the two versions of the API
