+++ apis/multiarch/common/const.go
@@ -1,0 +1,3 @@
+package common
+
+const SingletonResourceObjectName = "cluster"
