+++ apis/multiarch/v1alpha1/clusterPodPlacementConfig_conversion.go
@@ -30,7 +30,7 @@ func (src *ClusterPodPlacementConfig) ConvertTo(dstRaw conversion.Hub) error {
 	dst.ObjectMeta = src.ObjectMeta
 
 	// Spec
-	dst.Spec.LogVerbosity = multiarchv1beta1.LogVerbosityLevel(src.Spec.LogVerbosity)
+	dst.Spec.LogVerbosity = src.Spec.LogVerbosity
 	dst.Spec.NamespaceSelector = src.Spec.NamespaceSelector
 
 	// Status
@@ -53,7 +53,7 @@ func (dst *ClusterPodPlacementConfig) ConvertFrom(srcRaw conversion.Hub) error {
 	dst.ObjectMeta = src.ObjectMeta
 
 	// Spec
-	dst.Spec.LogVerbosity = LogVerbosityLevel(src.Spec.LogVerbosity)
+	dst.Spec.LogVerbosity = src.Spec.LogVerbosity
 	dst.Spec.NamespaceSelector = src.Spec.NamespaceSelector
 
 	// Status
