+++ apis/multiarch/v1alpha1/clusterpodplacementconfig_types.go
@@ -18,6 +18,8 @@ package v1alpha1
 
 import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+
+	"github.com/openshift/multiarch-tuning-operator/apis/multiarch/common"
 )
 
 // ClusterPodPlacementConfigSpec defines the desired state of ClusterPodPlacementConfig
@@ -27,7 +29,7 @@ type ClusterPodPlacementConfigSpec struct {
 	// Defaults to "Normal".
 	// +optional
 	// +kubebuilder:default=Normal
-	LogVerbosity LogVerbosityLevel `json:"logVerbosity,omitempty"`
+	LogVerbosity common.LogVerbosityLevel `json:"logVerbosity,omitempty"`
 
 	// NamespaceSelector filters the namespaces that the architecture aware pod placement can operate.
 	//
