+++ apis/multiarch/v1alpha1/groupversion_info.go
@@ -34,5 +34,3 @@ var (
 	// AddToScheme adds the types in this group-version to the given scheme.
 	AddToScheme = SchemeBuilder.AddToScheme
 )
-
-const SingletonResourceObjectName = "cluster"
