+++ controllers/operator/clusterpodplacementconfig_controller_test.go
@@ -19,7 +19,6 @@ package operator
 import (
 	"fmt"
 
-	"github.com/openshift/multiarch-tuning-operator/apis/multiarch/v1alpha1"
 	admissionv1 "k8s.io/api/admissionregistration/v1"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
@@ -30,6 +29,8 @@ import (
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 
+	"github.com/openshift/multiarch-tuning-operator/apis/multiarch/common"
+	"github.com/openshift/multiarch-tuning-operator/apis/multiarch/v1alpha1"
 	"github.com/openshift/multiarch-tuning-operator/pkg/utils"
 )
 
@@ -167,12 +168,12 @@ var _ = Describe("Controllers/ClusterPodPlacementConfig/ClusterPodPlacementConfi
 	When("The ClusterPodPlacementConfig", func() {
 		Context("is handling the lifecycle of the operand", func() {
 			BeforeEach(func() {
-				err := k8sClient.Create(ctx, newClusterPodPlacementConfig().WithName(v1alpha1.SingletonResourceObjectName).Build())
+				err := k8sClient.Create(ctx, newClusterPodPlacementConfig().WithName(common.SingletonResourceObjectName).Build())
 				Expect(err).NotTo(HaveOccurred(), "failed to create ClusterPodPlacementConfig", err)
 				validateReconcile()
 			})
 			AfterEach(func() {
-				err := k8sClient.Delete(ctx, newClusterPodPlacementConfig().WithName(v1alpha1.SingletonResourceObjectName).Build())
+				err := k8sClient.Delete(ctx, newClusterPodPlacementConfig().WithName(common.SingletonResourceObjectName).Build())
 				Expect(err).NotTo(HaveOccurred(), "failed to delete ClusterPodPlacementConfig", err)
 				validateDeletion()
 			})
@@ -300,13 +301,13 @@ var _ = Describe("Controllers/ClusterPodPlacementConfig/ClusterPodPlacementConfi
 				ppc2 := &v1alpha1.ClusterPodPlacementConfig{}
 				err := k8sClient.Get(ctx, crclient.ObjectKeyFromObject(&v1alpha1.ClusterPodPlacementConfig{
 					ObjectMeta: metav1.ObjectMeta{
-						Name:      v1alpha1.SingletonResourceObjectName,
+						Name:      common.SingletonResourceObjectName,
 						Namespace: utils.Namespace(),
 					},
 				}), ppc2)
 				Expect(err).NotTo(HaveOccurred(), "failed to get ClusterPodPlacementConfig", err)
 				// change the clusterpodplacementconfig's logLevel
-				ppc2.Spec.LogVerbosity = v1alpha1.LogVerbosityLevelTraceAll
+				ppc2.Spec.LogVerbosity = common.LogVerbosityLevelTraceAll
 				err = k8sClient.Update(ctx, ppc2)
 				Expect(err).NotTo(HaveOccurred(), "failed to update ClusterPodPlacementConfig", err)
 				Eventually(func(g Gomega) {
@@ -319,7 +320,7 @@ var _ = Describe("Controllers/ClusterPodPlacementConfig/ClusterPodPlacementConfi
 					}), &d)
 					g.Expect(err).NotTo(HaveOccurred(), "failed to get deployment "+PodPlacementControllerName, err)
 					g.Expect(d.Spec.Template.Spec.Containers[0].Args).To(ContainElement(
-						fmt.Sprintf("-zap-log-level=%d", v1alpha1.LogVerbosityLevelTraceAll.ToZapLevelInt())))
+						fmt.Sprintf("-zap-log-level=%d", common.LogVerbosityLevelTraceAll.ToZapLevelInt())))
 				}).Should(Succeed(), "the deployment "+PodPlacementControllerName+" should be updated")
 				Eventually(func(g Gomega) {
 					d := appsv1.Deployment{}
@@ -331,7 +332,7 @@ var _ = Describe("Controllers/ClusterPodPlacementConfig/ClusterPodPlacementConfi
 					}), &d)
 					g.Expect(err).NotTo(HaveOccurred(), "failed to get deployment "+PodPlacementWebhookName, err)
 					g.Expect(d.Spec.Template.Spec.Containers[0].Args).To(ContainElement(
-						fmt.Sprintf("-zap-log-level=%d", v1alpha1.LogVerbosityLevelTraceAll.ToZapLevelInt())))
+						fmt.Sprintf("-zap-log-level=%d", common.LogVerbosityLevelTraceAll.ToZapLevelInt())))
 				}).Should(Succeed(), "the deployment "+PodPlacementWebhookName+" should be updated")
 			})
 			It("Should sync the namespace selector", func() {
@@ -339,7 +340,7 @@ var _ = Describe("Controllers/ClusterPodPlacementConfig/ClusterPodPlacementConfi
 				ppc := &v1alpha1.ClusterPodPlacementConfig{}
 				err := k8sClient.Get(ctx, crclient.ObjectKeyFromObject(&v1alpha1.ClusterPodPlacementConfig{
 					ObjectMeta: metav1.ObjectMeta{
-						Name:      v1alpha1.SingletonResourceObjectName,
+						Name:      common.SingletonResourceObjectName,
 						Namespace: utils.Namespace(),
 					},
 				}), ppc)
@@ -387,7 +388,7 @@ func (p *clusterPodPlacementConfigFactory) WithNamespaceSelector(labelSelector *
 	return p
 }
 
-func (p *clusterPodPlacementConfigFactory) WithLogVerbosity(logVerbosity v1alpha1.LogVerbosityLevel) *clusterPodPlacementConfigFactory {
+func (p *clusterPodPlacementConfigFactory) WithLogVerbosity(logVerbosity common.LogVerbosityLevel) *clusterPodPlacementConfigFactory {
 	p.Spec.LogVerbosity = logVerbosity
 	return p
 }
